(fix) Parse "not created" status message
had to extract the parseStatus function to properly test it
