+++ index.js
@@ -37,7 +37,7 @@ util.inherits(Machine, EventEmitter);
 
 function _command(name, args, more) {
     more = more || [];
-    
+
     if(!args || (typeof args === 'function')) {
         args = [];
     }
@@ -109,7 +109,7 @@ function run(command, opts, cb) {
 
 
 Machine.prototype._run = function(command, cb) {
-    
+
     var self = this;
     if(self._runningCommand) {
         return cb(new Error('Already running command'));
@@ -137,7 +137,7 @@ Machine.prototype.sshConfig = function(cb) {
 
     this._run(command, function(err, out) {
         if(err) return cb(err);
-        
+
         var config = {};
         for(var key in SSH_CONFIG_MATCHERS) {
             config[key] = out.match(SSH_CONFIG_MATCHERS[key])[1];
@@ -153,26 +153,8 @@ Machine.prototype.status = function(cb) {
     this._run(command, function(err, out) {
         if(err) return cb(err);
 
-        var lines = out.split('\n').slice(2).reduce(function(prev, curr) {
-            if(prev.length > 0 && prev[prev.length - 1].length === 0) 
-                return prev;
-
-            prev.push(curr.trim());
-            return prev;
-        }, []);
-
-        lines.pop();
-        
-        var re = /(\S+)\s+(\S+)\s+\((\S+)\)/;
-
-        var statuses = {};
-        lines.forEach(function(line) {
-            var res = line.match(re);
-            statuses[res[1]] = {
-                status: res[2],
-                provider: res[3]
-            };
-        });
+        var statusParser = require("./parseStatus");
+        var statuses = statusParser(out);
 
         cb(null, statuses);
     });
@@ -279,7 +261,7 @@ module.exports.Machine = Machine;
 
 module.exports.globalStatus = function(args, cb) {
      cb = cb || args;
- 
+
      var command = _command('global-status', args);
      run(command, function(err, out) {
          if(err) return cb(err);
@@ -287,15 +269,15 @@ module.exports.globalStatus = function(args, cb) {
          var lines = out.split('\n').slice(2).reduce(function(prev, curr) {
              if(prev.length > 0 && prev[prev.length - 1].length === 0)
                  return prev;
- 
+
              prev.push(curr.trim());
              return prev;
          }, []);
- 
+
          lines.pop();
          if(/no active Vagrant environments/.test(lines[0]))
              lines = [];
- 
+
          var re = /(\S+)\s+(\S+)\s+(\S+)\s+(\S+)\s+(\S+)/;
          lines = lines.map(function(line) {
              var res = line.match(re);
@@ -319,4 +301,3 @@ module.exports.create = function(opts) {
 module.exports.version = function(cb) {
     run(_command('version'), cb);
 };
-
