+++ test/status
@@ -1,0 +1,8 @@
+Current machine states:
+
+my_server            running (docker)
+rethinkDB                 not created (virtualbox)
+
+This environment represents multiple VMs. The VMs are all listed
+above with their current state. For more information about a specific
+VM, run `vagrant status NAME`.
